Change color scheme into a component
